src/test/java/com/spartaglobal/ymao/sorters/BubbleSortTest.java
package com.spartaglobal.ymao.sorters;

import com.spartaglobal.ymao.model.sorters.BubbleSort;
import org.junit.Test;

import static org.junit.Assert.assertArrayEquals;

public class BubbleSortTest extends SortTest {

    @Override
    @Test
    public void testLarge() {
        initialiseLarge();
        assertArrayEquals(getResult(),
                BubbleSort.getInstance().sort(getInput()));
    }

    @Override
    @Test
    public void testZero() {
        initialiseZero();
        assertArrayEquals(getResult(),
                BubbleSort.getInstance().sort(getInput()));
    }

    @Override
    @Test
    public void testSorted() {
        initialiseLarge();
        assertArrayEquals(getResult(),
                BubbleSort.getInstance().sort(getResult()));
    }

}
